Move boards components into features
