+++ src/pages/boards/[id].tsx
@@ -3,10 +3,10 @@ import { useDispatch, useSelector } from "react-redux";
 import { useRouter } from "next/router";
 import styled from "styled-components";
 import Image from "next/image";
-import { Navbar } from "src/components/boards/navbar";
+import { Navbar } from "src/features/board/navbar";
 import { NewColumn } from "src/components/columns/new-column";
 import { ColumnList } from "src/components/columns/column-list";
-import { Subnav } from "src/components/boards/subnav";
+import { Subnav } from "src/features/board/subnav";
 import { imagePath } from "src/utils/image-paths";
 import { getBoardThunk, selectBoardById } from "src/redux/board.slice";
 import { useAuth } from "src/utils/use-auth";
